requirements.txt
Django>=5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-allauth
django-allauth[socialaccount]
django-widget-tweaks
Pillow
django-crispy-forms 
crispy-bootstrap5